requirements.txt
streamlit
Pillow
pyzbar
